Moving WeatherStats object into lib
